+++ Jellyfin.Plugin.Crunchyroll/Providers/CrunchyrollSeriesProvider.cs
@@ -46,7 +46,8 @@ public class CrunchyrollSeriesProvider : IRemoteMetadataProvider<Series, SeriesI
         var username = config?.Username;
         var password = config?.Password;
         var dockerContainerName = config?.DockerContainerName;
-        using var apiClient = new CrunchyrollApiClient(httpClient, _logger, locale, flareSolverrUrl, username, password, dockerContainerName);
+        var chromeCdpUrl = config?.ChromeCdpUrl;
+        using var apiClient = new CrunchyrollApiClient(httpClient, _logger, locale, flareSolverrUrl, username, password, dockerContainerName, chromeCdpUrl);
 
         string? crunchyrollId = info.GetProviderId("Crunchyroll");
         CrunchyrollSeries? series = null;
@@ -117,7 +118,8 @@ public class CrunchyrollSeriesProvider : IRemoteMetadataProvider<Series, SeriesI
         var username = config?.Username;
         var password = config?.Password;
         var dockerContainerName = config?.DockerContainerName;
-        using var apiClient = new CrunchyrollApiClient(httpClient, _logger, locale, flareSolverrUrl, username, password, dockerContainerName);
+        var chromeCdpUrl = config?.ChromeCdpUrl;
+        using var apiClient = new CrunchyrollApiClient(httpClient, _logger, locale, flareSolverrUrl, username, password, dockerContainerName, chromeCdpUrl);
 
         // Check if we have a Crunchyroll ID
         string? crunchyrollId = searchInfo.GetProviderId("Crunchyroll");